pyproject.toml
[tool.poetry]
name = "goit-web-homework-6"
version = "0.1.0"
description = ""
authors = ["Yuliia Didenko <yu_lyan@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
faker = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
